Setup menu flow and created cardgame view
